Minor fixes (SoftWipe analysis).

+++ src/libtrac.h
@@ -801,7 +801,7 @@ typedef struct {
   double met_dt_out;
 
   /*! Preload meteo data into disk cache (0=no, 1=yes). */
-  double met_cache;
+  int met_cache;
 
   /*! Isosurface parameter
      (0=none, 1=pressure, 2=density, 3=theta, 4=balloon). */
